BAU: Adjust the question request answer flow
This PR readjusts the question request answer flow to meet
the requirements of the sequeunce diagram. The /question endpoint
should only return a question at a time. And the /answer endpoint
should only send an answer one at a time and return a 200 OK status
when all questions are answered and the /question endpoint is hit
it should return a 204 NO_CONTENT indicating the all question have been
answered and there are no further questions.

Co-authored-by: aman.mohammend <aman.mohammed@digital.cabinet-office.gov.uk>