KBV-605: Fix test-logic of 500 no further question
The fix here is ensure that the RTQ endpoint (submitAnswers) is called when there are
no further questions. Initially the test implied it was the SAA (getQuestions)
endpoint that was called. Also the title was changed to reflect that
the questionResponse state is not an END state. The END state is usually
returned with the last batch of questions or it could be returned in
thin file situations when there are in sufficient questions in such state
when want to return with a 204 No content and not an error.