OJ-937: Assign kbvQuality to Evidence CheckDetail
NA Questions are either questions that have been deem not suitable for
KBV but are to be returned as LOW quality questions.
We currently have a mix of LOW and MEDIUM quality questions these question in the mapping have
value Zero. We are not sure if we need to log these in future, but we
know we want to VC to return LOW when these questions show up.

In a 3 out of 4 scenario, where for example 3 high-quality questions and
1 medium-quality are asked, and one of the question out of the initial 2 is incorrect,
it is not possible to tell whether the incorrect answer is a high-quality one
or a medium-quality one. The logic in EvidenceFactory.createKbvQualityEvidence
chooses a medium-quality over a high-quality question in the deciding the
last kbv-quality mapping of the 3 checkDetail object to include as part of
checkDetail[] node of the evidence

In a 3 out of 4 scenario, where for example 3 high-quality questions and
1 medium-quality are asked where both initial questions asked were answered
correctly. We know that the next question answered is wrong for there to
be additional question to be asked. In this case, kbvQuality can be mapped
exactly b'cos. The wrong question can be filtered from the checkDetail[]

see: https://github.com/alphagov/digital-identity-architecture/blob/main/rfc/0024-identity-vc-for-checks.md#user-content-knowledge-based-verification-kbv-check