OJ-2278: Add thin file for 3 question scenario
This occurs on the question answer handler, when the user
answer first 2 questions initially obtained from
the 3rd-party. On processing on the answer to these questions
on 3rd-party end one of the question is found to be wrong.
This results in a thin file situation b'cos the user needs
at least 2 more questions to pass KBV. But only now has one more on file