OJ-2520: handle resubmission of existing question
When a re-submission happens the question exists in “all QA pairs” and not in “QA pairs”
this means the question is a previously provided one in this situation continue the next question
should be provided. In a situation where there the question does not exist in both “all QA pairs” and “QA pairs”.
Or the “all QA pairs” and “QA pairs” are both empty then throw the illegal state exception.