OJ-2612: wrap common config with kbv config in QuestionHandler
Wrap common lib config and expose a new method getKbvQuestionStrategy
so it can be used else where. other common lib configuration methods
are to be accessed with get().<method names>

OJ-2612: use Kbv config strategy for question handler

The verification score was harded to 2. This commit
uses the value from the evidenceRequest of the sessionItem
which default to 2 if no value was written to storage