OJ-2643: Introduce KbvStrategyParser
In order to CheckDetails creation in the EvidenceFactory.
The KbvStrategyParser will be used in conjunction with the Strategy class.
The KbvStrategyParser will take as input any of the two known kbv question strategy i.e

The existing Medium confidence = "3 out of 4 Prioritised"
The new low confidence         = "2 out of 3 Prioritised"

KbvStrategyParser.parse() method would return the minimum and maximum
number of questions need for a successful KBV pass the getMin() and the
getMax(). i.e for the kbvStrategy

The existing Medium confidence = "  3 out of  4 Prioritised"
                       min()  =  3  and max() = 4
The new low confidence         = "2 out of 3 Prioritised"
                       min() = 2 and max() = 3

This would allow the algorithm in EvidenceFactory to change mode
which ever strategy is choosen

 private boolean hasPassedWithOneIncorrectAnswer(KBVItem kbvItem) {
        return Objects.nonNull(kbvItem.getQuestionAnswerResultSummary())
                && kbvItem.getQuestionAnswerResultSummary().getQuestionsAsked() == 4
                && kbvItem.getQuestionAnswerResultSummary().getAnsweredCorrect() == 3
Becomes

 private boolean hasPassedWithOneIncorrectAnswer(KBVItem kbvItem) {
        return Objects.nonNull(kbvItem.getQuestionAnswerResultSummary())
                && kbvItem.getQuestionAnswerResultSummary().getQuestionsAsked() == strategy.max()
                && kbvItem.getQuestionAnswerResultSummary().getAnsweredCorrect() == strategy.min()