OJ-2643: Get CheckDetail algorithm working for 2 of 3
Modified allQuestionBatchSizesMatch by renaming to
isQuestionReceivedBatchCountEqualTo

This method takes in an argument of 2. This means exactly
2 requests to the 3rd-party. i.e isQuestionReceivedBatchCountEqualTo(2)

In 3 out of 4, it called after the check on if hasPassedWithOneIncorrectAnswer
is true, it also means one of the question from the first batch of two Questions
must have been wrong. B'cos the next of set of question from the 3rd-party must
have been a batch with 2 questions (making a total of 2 batches that make 4 questions)

This lead to executing `createCheckDetailsBySortingOnKbvQuality`
which require since we don't know exactly which of the first 2 question
was wrong.

NB: If two initial question where correct provide hasPassedWithOneIncorrectAnswer
has pass initially. If the next question was wrong then 3 batch would be required
not 2

In 2 out of 3 scenario the only time there will be 2 batches is if the
one of the first two question was wrong. In which case we also don't know
which particular one was wrong and there requests and therefore
the code branches to `createCheckDetailsBySortingOnKbvQuality`