OJ-2778: Resolving linting issue
Check Fn::If has a path that cannot be reached] (['Fn::If', 1] is not reachable.
When setting condition 'IsDevEnvironment' to True. Where existing status for condition 'IsNotDevEnvironment' is True)

The basically saying the condition `IsDevEnvironment`` is never used because of the condition IsNotDevEnvironment
applied when the resource is defined therefore the IsDevEnvironment can be removed entirely.