feat: use isTokenValidWithinThreshold IdentityIQWrapper
refs: OJ-2842

OJ-2842: create IdentityIQWrapper

The IdentityIQWrapper class, wrap IdentityIQWebServiceSoap class
It will manage the token refresh on expiry of a long running
call to SAA and RTQ endpoints

Creation of SoapToken, SoapRetriever and binding to the HeaderHandlerResolver
is delayed until the `.get`of the suppliers are called for instance
the following suppliers are specified in the ServiceFactory.

    private Supplier<KBVClientFactory> getKbvClientFactory() {
        return () ->
                new KBVClientFactory(
                        new IdentityIQWebService(),
                        new HeaderHandlerResolver(new HeaderHandler(getSoapTokenRetriever().get())),
                        getConfigurationService());
    }

    private Supplier<SoapTokenRetriever> getSoapTokenRetriever() {
        return () ->
                new SoapTokenRetriever(
                        new SoapToken(
                                APPLICATION, true, new TokenService(), getConfigurationService()));
    }

But this is not called until, the first time the IdentityIQWrapper is created through
the `refreshClient()` method in the constructor.

    private void refreshClient() {
        this.soapTokenRetriever = soapTokenRetrieverSupplier.get();
        this.soapToken = soapTokenRetriever.getSoapToken();
        this.identityIQWebServiceSoap = kbvClientFactorySupplier.get().createClient();
    }

The call kbvClientFactorySupplier.get().createClient(); will ensure the HeaderHandlerResolver
Binding takes place

The `refreshSoapToken()` checks if the current soapToken has expired. If it has
It calls `refreshClient()` which will generate a new token reinitialize the client
by ensuring the HeaderHandlerResolver binding takes place